[FEATURE] Add file-based log service
This commit adds a file-based log service. It can be called as follows:

LogService::log("The log message", LogService::WARNING);

The second parameter provides the severity level of the log message. It
respects the minimum log level which can be defined as
`MINIMUM_LOG_LEVEL` in your custom `.env` file.

Related: #29