[FEATURE] Provide a scheduler service to run tasks at a given time
This commit provides a scheduler service which can be used to schedule
and execute tasks at a given time in the future. For this, the
appropriate class should implement the TaskInterface. It can then be
executed using the new `scheduler:run` console command. Check README for
further instructions or run `./console.php help scheduler:run`.

Related: #12